requirements.txt
avro-python3==1.9.1
kafka-python==1.4.7
requests==2.22.0
urllib3==1.25.6
